start splitting huge file into chuncks
