+++ backend/app/routes.py
@@ -6,7 +6,7 @@ from app.db import pool
 from app.nodes.nearby.here import get_here_nodes_within
 from app.nodes.byID.here import get_here_node
 from app.nodes.byID.centreline import get_centreline_node
-from app.get_travel_time import get_travel_time
+from app.travel_times.get_travel_time import get_travel_time
 from app.links.here import get_here_links
 from app.links.centreline import get_centreline_links
 from app.getGitHash import getGitHash
