+++ backend/app/travel_times/cache.py
@@ -1,0 +1,29 @@
+from app.db import pool
+from app.getGitHash import getGitHash
+
+def checkCache(uri):
+    query = f'''
+        SELECT results
+        FROM nwessel.cached_travel_times
+        WHERE uri_string = %(uri)s AND commit_hash = %(hash)s
+    '''
+    with pool.connection() as connection:
+        with connection.cursor() as cursor:
+            try:
+                cursor.execute(query, {'uri': uri, 'hash': getGitHash()})
+                for (record,) in cursor: # will skip if no records
+                    return record # there could only be one
+            except:
+                pass
+
+def cacheAndReturn(obj,uri):
+    query = f'''
+        INSERT INTO nwessel.cached_travel_times (uri_string, commit_hash, results)
+        VALUES (%(uri)s, %(hash)s, %(results)s)
+    '''
+    with pool.connection() as connection:
+        with connection.cursor() as cursor:
+            try:
+                cursor.execute(query, {'uri': uri, 'hash': getGitHash(), 'results': json.dumps(obj)})
+            finally:
+                return obj
