+++ backend/app/travel_times/daily_aggregation.py
@@ -1,0 +1,12 @@
+import numpy
+
+# the way we currently do it
+def mean_daily_mean(obs):
+    # group the observations by date
+    dates = {}
+    for (dt,tt) in obs:
+        dates[dt] = [tt] if not dt in dates else dates[dt] + [tt]
+    # take the daily averages
+    daily_means = [ numpy.mean(times) for times in dates.values() ]
+    # average the days together
+    return numpy.mean(daily_means)
