+++ backend/app/travel_times/get_travel_time.py
@@ -4,53 +4,14 @@ from app.db import pool
 from app.links.here import get_here_links
 from app.nodes.byID.here import get_here_node
 from app.hereMapVersions import selectMapVersions
-from app.getGitHash import getGitHash
 from traveltimetools.utils import timeFormats
+from app.travel_times.cache import checkCache, cacheAndReturn
+from app.travel_times.bootstrap import bootstrap
+from app.travel_times.daily_aggregation import mean_daily_mean
 from haversine import haversine, Unit
 from functools import reduce
-import numpy
-import math
 import pandas
-import random
 import json
-
-# the way we currently do it
-def mean_daily_mean(obs):
-    # group the observations by date
-    dates = {}
-    for (dt,tt) in obs:
-        dates[dt] = [tt] if not dt in dates else dates[dt] + [tt]
-    # take the daily averages
-    daily_means = [ numpy.mean(times) for times in dates.values() ]
-    # average the days together
-    return numpy.mean(daily_means)
-
-def checkCache(uri):
-    query = f'''
-        SELECT results
-        FROM nwessel.cached_travel_times
-        WHERE uri_string = %(uri)s AND commit_hash = %(hash)s
-    '''
-    with pool.connection() as connection:
-        with connection.cursor() as cursor:
-            try:
-                cursor.execute(query, {'uri': uri, 'hash': getGitHash()})
-                for (record,) in cursor: # will skip if no records
-                    return record # there could only be one
-            except:
-                pass
-
-def cacheAndReturn(obj,uri):
-    query = f'''
-        INSERT INTO nwessel.cached_travel_times (uri_string, commit_hash, results)
-        VALUES (%(uri)s, %(hash)s, %(results)s)
-    '''
-    with pool.connection() as connection:
-        with connection.cursor() as cursor:
-            try:
-                cursor.execute(query, {'uri': uri, 'hash': getGitHash(), 'results': json.dumps(obj)})
-            finally:
-                return obj
 
 def addLinkLengths(a,b):
     return a['length_m'] + b['length_m']
@@ -218,20 +179,7 @@ def get_travel_time(start_node, end_node, start_time, end_time, start_date, end_
 
     tt_seconds = mean_daily_mean(sample)
 
-    reported_intervals = None
-    if len(sample) > 1:
-        # bootstrap for synthetic sample distribution
-        sample_distribution = []
-        for i in range(0,100):
-            bootstrap_sample = random.choices( sample, k = len(sample) )
-            sample_distribution.append( mean_daily_mean(bootstrap_sample) )
-        p95lower, p95upper = numpy.percentile(sample_distribution, [2.5, 97.5])
-        reported_intervals = {
-            'p=0.95': {
-                'lower': timeFormats(p95lower,1),
-                'upper': timeFormats(p95upper,1)
-            }
-        }
+    reported_intervals = bootstrap(sample)
 
     return cacheAndReturn({
         'results': {
